content/media/webaudio/test/test_mediaStreamAudioSourceNode.html

<!DOCTYPE HTML>
<html>
<meta charset="utf-8">
<head>
  <title>Test MediaStreamAudioSourceNode processing is correct</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
var gTest;

// This makes loading audio *much* easier.
var url = "sin-440-1024smpl-48k.wav";

// First, decode the expected buffer using decodeAudioData
var xhr = new XMLHttpRequest();
xhr.open("GET", url, true);
xhr.responseType = "arraybuffer";
xhr.onload = function() {
  (new AudioContext()).decodeAudioData(xhr.response, function(decodedBuffer) {
    gTest = {
      length: 1024,
      numberOfChannels: decodedBuffer.numberOfChannels,
      skipOfflineContextTests: true,
      createGraphAsync: function(context, callback) {
        var audio = document.createElement('audio');
        document.body.appendChild(audio);
        audio.preload = 'auto';
        audio.src = url;
        audio.load();
        var stream = audio.mozCaptureStream();
        var source = context.createMediaStreamSource(stream);
        callback(source);
        audio.play();
      },
      createExpectedBuffers: function(context) {
        var buffer = context.createBuffer(decodedBuffer.numberOfChannels,
                                          1024,
                                          context.sampleRate);
        for (var i = 0; i < buffer.numberOfChannels; ++i) {
          for (var j = 0; j < 1024; j++) {
            buffer.getChannelData(i)[j] = decodedBuffer.getChannelData(i)[j];
          }
        }

        return buffer;
      }
    };

    runTest();
  }, function() {
    ok(false, "Decoding the data failed");
  });
};
xhr.send();
</script>
</pre>
</body>
</html>
